projects/hey-lantern/software/pyproject.toml
[project]
name = "hey-lantern"
version = "0.1.0"
description = "Voice assistant POC for Hey Lantern"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "openwakeword>=0.6.0",
    "sounddevice>=0.5.0",
    "soundfile>=0.12.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
]
